demo/requirements.txt
# Demo Requirements

# UI
streamlit>=1.31.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Database
duckdb>=0.9.0

# Optimization
pulp>=2.7.0  # Linear programming solver

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
